backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
playwright==1.44.0
python-multipart==0.0.9
pydantic==2.7.1
anthropic==0.28.0
reportlab==4.2.0
pdfplumber==0.11.0
python-docx==1.1.2
